refactor: moved Ethereum client to pkg/chains/evm
